Moves validation and normalization mixins to utils
(Part of a patchset focusing on a step data conversion to dataclasses.)

These will be used from other modules, and having them in tmt.base would
cause a nasty and hard to break cyclic imports.

To improve the reusability, normalization mixin is split to a
source-agnostic class which can be used with any mapping-like source,
and a thin child class serving as a glue for base classes and their fmf
sources.

Logging has been improved, too, by adding a pre/post dump of fields.
